requirements.txt
absl-py==2.1.0
attrs==23.2.0
cffi==1.16.0
contourpy==1.2.1
cycler==0.12.1
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
imageio==2.34.2
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.1
mediapipe==0.10.14
ml-dtypes==0.4.0
mpmath==1.3.0
networkx==3.3
numpy==2.0.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pillow==10.4.0
protobuf==4.25.3
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scikit-image==0.24.0
scipy==1.14.0
sentencepiece==0.2.0
six==1.16.0
sounddevice==0.4.7
sympy==1.13.0
tifffile==2024.7.2
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
triton==2.3.1
typing_extensions==4.12.2
